reorganize modules structure
